fix: removed re-rendering between user rating, separated logic into wine cards
