+++ client/src/components/Results.tsx
@@ -6,19 +6,13 @@ import {
   Flex,
   Grid,
   Heading,
-  Image,
-  Text,
-  VStack,
 } from '@chakra-ui/react';
 import { fetchFilteredWines } from '../services/wineService';
 import { fetchFavouriteWines } from '../services/favouritesService';
 import { submitRating, deleteRating } from '../services/ratingService';
 import { useAuth } from '../context/useAuth';
 import type { Wine } from '../types/wine';
-import { motion } from 'framer-motion';
-import StarRating from './StarRating';
-
-const BASE_URL = import.meta.env.VITE_API_BASE_URL;
+import WineCard from './WineCard';
 
 function Results({ favourites = false }: { favourites?: boolean }) {
   const [wines, setWines] = useState<Wine[]>([]);
@@ -33,8 +27,6 @@ function Results({ favourites = false }: { favourites?: boolean }) {
   const pairing = query.get('pairing') || '';
   const min = query.get('min');
   const max = query.get('max');
-
-  const MotionBox = motion.create(Box);
 
   useEffect(() => {
     // Delay initial render to allow fetch to complete
@@ -207,79 +199,13 @@ function Results({ favourites = false }: { favourites?: boolean }) {
         pb={12}
       >
         {shouldRender && wines.map((wine, index) => (
-          <MotionBox
+          <WineCard
+            index={index}
             key={wine.id}
-            initial={{ opacity: 0, y: 20 }}
-            animate={{ opacity: 1, y: 0 }}
-            transition={{ delay: index * 0.6, duration: 0.8 }}
-            position="relative"
-            w="260px"
-            bg="white"
-            borderRadius="16px"
-            boxShadow="0 4px 14px rgba(0, 0, 0, 0.08)"
-            cursor="pointer"
-            p="1rem 1rem 1.5rem"
-            transitionProperty="transform, box-shadow"
-            transitionDuration="0.3s"
-            transitionTimingFunction="ease"
-            onClick={() => handleSelect(wine)}
-            _hover={{
-              transform: "scale(1.03)",
-              boxShadow: "0 6px 20px rgba(123, 46, 90, 0.2)",
-              zIndex: 2,
-              "& .wine-image-wrapper": {
-                top: "-60px",
-              },
-              "& .wine-image": {
-                transform: "scale(1.12)",
-              },
-            }}
-          >
-            <Box
-              className="wine-image-wrapper"
-              position="relative"
-              top="-50px"
-              display="flex"
-              justifyContent="center"
-              zIndex={1}
-              transition="top 0.3s ease"
-            >
-              <Image
-                className="wine-image"
-                src={`${BASE_URL}/${wine.image_url}`}
-                alt={wine.name}
-                w="100px"
-                h="auto"
-                objectFit="contain"
-                transition="transform 0.3s ease"
-                zIndex={3}
-              />
-            </Box>
-            
-            <VStack
-              spacing={2}
-              pt={2}
-              textAlign="center"
-            >
-              <Text
-                fontWeight="600"
-                color="brand.primary"
-                fontSize="md"
-              >
-                {wine.name}
-              </Text>
-              <Text fontFamily="body" fontSize="sm" color="gray.600">
-                <Text as="span" fontWeight="bold">Country:</Text> {wine.country}
-              </Text>
-              <Text fontFamily="body" fontSize="sm" color="gray.600">
-                <Text as="span" fontWeight="bold">Grape:</Text> {wine.grape}
-              </Text>
-              <Text fontFamily="body" fontSize="sm" color="gray.600">
-                <Text as="span" fontWeight="bold">Price:</Text> ${wine.price}
-              </Text>
-              <StarRating rating={wine.ratings?.[0]?.score ?? 0} onRate={user ? (value) => handleRating(wine.id, value) : undefined} ></StarRating>
-            </VStack>
-          </MotionBox>
+            wine={wine}
+            onRate={(value) => handleRating(wine.id, value)}
+            onSelect={() => handleSelect(wine)}
+          />
         ))}
       </Grid>
     </Box>
