+++ server/tsconfig.json
@@ -2,7 +2,7 @@
   "compilerOptions": {
     "target": "ES2020",
     "module": "commonjs",
-    "rootDir": "src",
+    "rootDir": "./",
     "outDir": "dist",
     "strict": true,
     "esModuleInterop": true,
